app/requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coverage==7.6.12
cryptography==44.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
